test(e2e): migrate Image Snap smoke to Appium
